types for all the actions - need to work on bringing them into their own file
