+++ app/src/index.js
@@ -2,13 +2,11 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 // import './index.css';
 import App from './components/App';
-import App2 from './components/App2';
 import * as serviceWorker from './serviceWorker';
 
 ReactDOM.render(
   <React.StrictMode>
-    <App2 />
-    {/* <App /> */}
+    <App />
   </React.StrictMode>,
   document.getElementById('root')
 );
